# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillScope – AI-Powered Developer Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /github/profile:
    get:
      operationId: getGithubProfile
      tags: [github]
      summary: Get connected GitHub analytics profile
      responses:
        "200":
          description: GitHub profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubProfile"
        "404":
          description: Not connected
  /github/connect:
    post:
      operationId: connectGithub
      tags: [github]
      summary: Connect or refresh GitHub profile by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubConnectInput"
      responses:
        "200":
          description: GitHub profile connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubProfile"
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get detailed project intelligence for a repo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found

  # ─── Resumes ─────────────────────────────────────────────────────────────────
...
  /public/profile/{username}:
    get:
      operationId: getPublicProfile
      tags: [public]
      summary: Get public developer profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Not found
components:
  schemas:
...
    GithubProfile:
      type: object
      required: [id, username, totalRepos, stars, forks, followers, following, commits, pullRequests, issues, score]
      properties:
        id:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        totalRepos:
          type: integer
        stars:
          type: integer
        forks:
          type: integer
        followers:
          type: integer
        following:
          type: integer
        commits:
          type: integer
        pullRequests:
          type: integer
        issues:
          type: integer
        score:
          type: integer
        topLanguage:
          type: ["string", "null"]
        weeklyHours:
          type: ["number", "null"]
        trendingRepo:
          type: ["string", "null"]
    GithubConnectInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
...
    CodingProfile:
      type: object
      required: [id, platform, username, problemsSolved, rating, rank]
      properties:
        id:
          type: integer
        platform:
          type: string
          enum: [leetcode, codeforces, codechef, hackerrank, geeksforgeeks, atcoder]
        username:
          type: string
        problemsSolved:
          type: integer
        rating:
          type: ["integer", "null"]
        rank:
          type: ["string", "null"]
        streak:
          type: ["integer", "null"]
        easyCount:
          type: ["integer", "null"]
        mediumCount:
          type: ["integer", "null"]
        hardCount:
          type: ["integer", "null"]
        badges:
          type: integer
        lastSynced:
          type: ["string", "null"]
...
    Skill:
      type: object
      required: [id, name, category, score, confidence, color]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        score:
          type: integer
        confidence:
          type: string
          enum: [low, medium, high]
        improvementSuggestion:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
...
    Project:
      type: object
      required: [id, name, language, stars, difficulty, score]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        stars:
          type: integer
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        score:
          type: integer
        techStack:
          type: array
          items:
            type: string
        htmlUrl:
          type: ["string", "null"]
    ProjectDetail:
      type: object
      required: [id, name, score, difficulty, codeComplexity, maintainabilityIndex, documentationScore, readabilityScore, testingScore, securityScore]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        architecture:
          type: ["string", "null"]
        techStack:
          type: array
          items:
            type: string
        htmlUrl:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        language:
          type: ["string", "null"]
        score:
          type: integer
        codeComplexity:
          type: integer
        maintainabilityIndex:
          type: integer
        documentationScore:
          type: integer
        readabilityScore:
          type: integer
        testingScore:
          type: integer
        securityScore:
          type: integer
        performanceSuggestions:
          type: array
          items:
            type: string
        designPatterns:
          type: array
          items:
            type: string

    # ─── Resumes ────────────────────────────────────────────────────────────────
...
    Badge:
      type: object
      required: [id, name, description, icon, earnedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        earnedAt:
          type: string
        rarity:
          type: string
          enum: [common, rare, epic, legendary]
...
    PublicProfile:
      type: object
      required: [username, name, overallScore, skills, projects, githubProfile]
      properties:
        username:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        overallScore:
          type: integer
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        githubProfile:
          $ref: "#/components/schemas/GithubProfile"
        codingProfiles:
          type: array
          items:
            $ref: "#/components/schemas/CodingProfile"
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"